tests/api/requirements.txt
# HolmOS API Test Dependencies

# Core testing framework
pytest>=7.0.0

# HTTP client
requests>=2.28.0

# YAML parsing for services.yaml
PyYAML>=6.0

# HTML report generation
pytest-html>=3.2.0

# JSON report generation
pytest-json-report>=1.5.0

# Timeout handling
pytest-timeout>=2.1.0

# Parallel test execution (optional)
pytest-xdist>=3.0.0

# Code coverage (optional)
pytest-cov>=4.0.0

# WebSocket testing (optional)
websocket-client>=1.4.0
